Bring enhancement from previous PreferencesModal
